requirements.txt
altair
annotated-types
anthropic
anyio
attrs
beautifulsoup4
blinker
bs4
cachetools
certifi
charset-normalizer
click
distro
dnspython
duckduckgo-search
exceptiongroup
filelock
fsspec
gitdb
gitpython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
groq
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
huggingface-hub
idna
jinja2
jiter
joblib
jsonschema
jsonschema-specifications
markdown-it-py
markupsafe
mdurl
mpmath
narwhals
networkx
numpy
ollama
openai
packaging
pandas
pillow
primp
proto-plus
protobuf
psutil
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pydeck
pygments
pymongo
pyparsing
pypdf2
pyreqwest-impersonate
python-dateutil
python-dotenv
pytz
pyyaml
referencing
regex
requests
rich
rpds-py
rsa
sacremoses
safetensors
scrape-web
setuptools
six
smmap
sniffio
soupsieve
streamlit
sympy
tavily-python
tenacity
tiktoken
tokenizers
toml
toolz
torch
tornado
tqdm
transformers
typing-extensions
tzdata
uritemplate
urllib3
